+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Initial Release
+- Added Muify identity attribute analyzer diagnostics MUIFY01 and MUIFY02.